what i have been trying to do since forever: add gaussdistr to funparser

+++ RainCode/src/cparse/GaussDistr.java
@@ -10,7 +10,7 @@ public class GaussDistr {
         random = new Random(seed);
     }
 
-    public double[] distribute(int x, int y, int z, int count) {
+    public double[] distribute(double x, double y, double z, long count) {
         double[] ret = new double[3];
         ret [0] = x + random.nextGaussian()*count;
         ret [1] = y + random.nextGaussian()*count;
